useRestaurantとPagesコンポーネントの処理の責任範囲が適切でなかったのでロジック見直し
